Решена проблема с удалением обработчика загрузки новых комментариев
