+++ js/elements-render/big-picture-modal-render.js
@@ -1,5 +1,3 @@
-import { commentsPartRender, resetCommentsVariables } from './comments-render.js';
-
 const bigPictureModal = document.querySelector('.big-picture');
 const bigPictureCloseButton = bigPictureModal.querySelector('.big-picture__cancel');
 const bigPictureImg = bigPictureModal.querySelector('.big-picture__img img');
@@ -8,6 +6,12 @@ const bigPictureDescription = bigPictureModal.querySelector('.social__caption');
 const bigPictureTotalCommentsCount = bigPictureModal.querySelector('.social__comment-total-count');
 const bigPictureCommentsList = document.querySelector('.social__comments');
 const bigPictureCommentsLoader = document.querySelector('.comments-loader');
+const commentTemplateItem = document.querySelector('#comment').content.querySelector('.social__comment');
+const bigPictureCommentsShownCount = document.querySelector('.social__comment-shown-count');
+
+const PART_OF_COMMENTS = 5;
+let comments = [];
+let renderedComments = 0;
 
 function escapeKeydown(evt) {
   if (evt.key === 'Escape') {
@@ -27,12 +31,19 @@ function bigPictureCloseButtonDown(evt) {
   }
 }
 
+function loadCommentsListener(evt) {
+  if (evt.target === bigPictureCommentsLoader) {
+    commentsPartRender(comments);
+  }
+}
+
 function bigPictureModalOpener() {
   bigPictureModal.classList.remove('hidden');
   document.body.classList.add('modal-open');
 
   document.addEventListener('keydown', escapeKeydown);
   document.addEventListener('click', mouseButtonDown);
+  bigPictureCommentsLoader.addEventListener('click', loadCommentsListener);
   bigPictureCloseButton.addEventListener('click', bigPictureCloseButtonDown);
 }
 
@@ -42,13 +53,55 @@ function bigPictureModalCloser() {
 
   document.removeEventListener('keydown', escapeKeydown);
   document.removeEventListener('click', mouseButtonDown);
+  bigPictureCommentsLoader.removeEventListener('click', loadCommentsListener);
   bigPictureCloseButton.removeEventListener('click', bigPictureCloseButtonDown);
 
   if (bigPictureCommentsLoader.classList.contains('hidden')) {
     bigPictureCommentsLoader.classList.remove('hidden');
   }
 
-  resetCommentsVariables();
+  renderedComments = 0;
+}
+
+function commentRender(comment) {
+  const commentsItem = commentTemplateItem.cloneNode(true);
+
+  commentsItem.querySelector('.social__picture').src = comment.avatar;
+  commentsItem.querySelector('.social__picture').alt = comment.name;
+  commentsItem.querySelector('.social__text').textContent = comment.message;
+
+  return commentsItem;
+}
+
+function commentsPartRender(commentsArray) {
+  if(commentsArray) {
+    comments = commentsArray;
+  }
+
+  let commentsToRenderCount = Math.min(renderedComments + PART_OF_COMMENTS, comments.length);
+
+  if (commentsToRenderCount >= comments.length) {
+    commentsToRenderCount = comments.length;
+  }
+
+  const commentsFragment = document.createDocumentFragment();
+
+  for (let i = renderedComments; i < commentsToRenderCount; i++) {
+    const comment = commentRender(comments[i]);
+    commentsFragment.appendChild(comment);
+  }
+
+  bigPictureCommentsList.appendChild(commentsFragment);
+
+  renderedComments = commentsToRenderCount;
+
+  if (commentsToRenderCount >= comments.length) {
+    bigPictureCommentsLoader.classList.add('hidden');
+  } else {
+    bigPictureCommentsLoader.classList.remove('hidden');
+  }
+
+  bigPictureCommentsShownCount.textContent = renderedComments;
 }
 
 function bigPictureModalRender(item) {
@@ -61,9 +114,7 @@ function bigPictureModalRender(item) {
 
   bigPictureCommentsList.replaceChildren();
 
-  commentsPartRender(item.comments, bigPictureCommentsList);
-
-  bigPictureCommentsLoader.addEventListener('click', () => commentsPartRender(item.comments, bigPictureCommentsList));
+  commentsPartRender(item.comments);
 }
 
 export { bigPictureModalRender };
